Filtered words feature - removes stop words from trending words view
