+++ news/utils/__init__.py
@@ -1,0 +1,5 @@
+# coding=utf-8
+import logging
+
+__author__ = 'ilov3'
+logger = logging.getLogger(__name__)
